[Refactor] Optimise header fanout P2
