+++ be/src/agent/agent_task.cpp
@@ -27,6 +27,7 @@
 #include "gutil/strings/join.h"
 #include "io/io_profiler.h"
 #include "runtime/current_thread.h"
+#include "runtime/exec_env.h"
 #include "runtime/snapshot_loader.h"
 #include "storage/lake/replication_txn_manager.h"
 #include "storage/lake/schema_change.h"
