+++ be/src/agent/heartbeat_server.h
@@ -42,7 +42,7 @@
 #include "gen_cpp/HeartbeatService.h"
 #include "gen_cpp/Status_types.h"
 #include "gutil/macros.h"
-#include "runtime/exec_env.h"
+#include "runtime/runtime_fwd.h"
 #include "storage/olap_define.h"
 #include "thrift/transport/TTransportUtils.h"
