+++ be/src/connector/file_chunk_sink.cpp
@@ -26,6 +26,7 @@
 #include "formats/parquet/parquet_file_writer.h"
 #include "formats/utils.h"
 #include "fs/fs_factory.h"
+#include "runtime/runtime_state.h"
 #include "utils.h"
 
 namespace starrocks::connector {
