+++ be/src/connector/file_chunk_sink.h
@@ -19,14 +19,11 @@
 #include <boost/thread/future.hpp>
 #include <future>
 
-#include "column/chunk.h"
 #include "common/status.h"
 #include "connector/connector.h"
 #include "connector_chunk_sink.h"
 #include "formats/column_evaluator.h"
 #include "formats/file_writer.h"
-#include "fs/fs.h"
-#include "runtime/runtime_state.h"
 #include "util/priority_thread_pool.hpp"
 #include "utils.h"
