+++ be/src/connector/partition_chunk_writer.cpp
@@ -23,7 +23,9 @@
 #include "connector/sink_memory_manager.h"
 #include "exec/pipeline/fragment_context.h"
 #include "formats/file_writer.h"
+#include "fs/fs.h"
 #include "runtime/descriptors.h"
+#include "runtime/exec_env.h"
 #include "runtime/runtime_state.h"
 #include "storage/chunk_helper.h"
 #include "storage/convert_helper.h"
