+++ be/src/exec/pipeline/exchange/local_exchange.h
@@ -23,10 +23,10 @@
 #include "exec/pipeline/exchange/local_exchange_source_operator.h"
 #include "exec/pipeline/exchange/shuffler.h"
 #include "exprs/expr_context.h"
+#include "runtime/runtime_state.h"
 
 namespace starrocks {
 class ExprContext;
-class RuntimeState;
 
 namespace pipeline {
