+++ be/src/formats/parquet/parquet_file_writer.h
@@ -56,9 +56,9 @@
 #include "formats/parquet/chunk_writer.h"
 #include "formats/parquet/file_writer.h"
 #include "formats/utils.h"
-#include "fs/fs.h"
+#include "fs/fs_fwd.h"
 #include "gen_cpp/Types_types.h"
-#include "runtime/runtime_state.h"
+#include "runtime/runtime_fwd.h"
 #include "types/type_descriptor.h"
 #include "util/priority_thread_pool.hpp"
 
@@ -71,7 +71,6 @@ namespace starrocks {
 class Chunk;
 class FileSystem;
 class PriorityThreadPool;
-class RuntimeState;
 
 namespace parquet {
 class ChunkWriter;
