+++ be/src/runtime/runtime_filter/runtime_filter_probe.cpp
@@ -27,6 +27,7 @@
 #include "gutil/strings/substitute.h"
 #include "runtime/exec_env.h"
 #include "runtime/runtime_filter_cache.h"
+#include "runtime/runtime_state.h"
 
 namespace starrocks {
 DEFINE_FAIL_POINT(global_runtime_filter_sync_B);
