+++ be/src/runtime/runtime_filter/runtime_filter_probe.h
@@ -34,7 +34,7 @@
 #include "gen_cpp/RuntimeFilter_types.h"
 #include "runtime/runtime_filter.h"
 #include "runtime/runtime_filter_layout.h"
-#include "runtime/runtime_state.h"
+#include "runtime/runtime_fwd.h"
 #include "types/logical_type.h"
 
 namespace starrocks {
